+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/BaseSearchSuggestionsCallback.kt
@@ -12,7 +12,7 @@ interface BaseSearchSelectionCallback : BaseSearchSuggestionsCallback {
 
     fun onResult(suggestion: BaseSearchSuggestion, result: BaseSearchResult, responseInfo: BaseResponseInfo)
 
-    fun onCategoryResult(
+    fun onResults(
         suggestion: BaseSearchSuggestion,
         results: List<BaseSearchResult>,
         responseInfo: BaseResponseInfo