+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/ForwardGeocodingBatchResolvingJavaExampleActivity.java
@@ -16,7 +16,6 @@ import com.mapbox.search.SearchSelectionCallback;
 import com.mapbox.search.common.AsyncOperationTask;
 import com.mapbox.search.result.SearchResult;
 import com.mapbox.search.result.SearchSuggestion;
-import com.mapbox.search.sample.BuildConfig;
 import com.mapbox.search.sample.R;
 
 import java.util.List;
@@ -44,7 +43,7 @@ public class ForwardGeocodingBatchResolvingJavaExampleActivity extends AppCompat
         }
 
         @Override
-        public void onCategoryResult(@NonNull SearchSuggestion suggestion, @NonNull List<SearchResult> results, @NonNull ResponseInfo responseInfo) {
+        public void onResults(@NonNull SearchSuggestion suggestion, @NonNull List<SearchResult> results, @NonNull ResponseInfo responseInfo) {
             Log.i("SearchApiExample", "Category search results: " + results);
         }