+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/ForwardGeocodingBatchResolvingKotlinExampleActivity.kt
@@ -38,7 +38,7 @@ class ForwardGeocodingBatchResolvingKotlinExampleActivity : Activity() {
             Log.i("SearchApiExample", "Search result: $result")
         }
 
-        override fun onCategoryResult(
+        override fun onResults(
             suggestion: SearchSuggestion,
             results: List<SearchResult>,
             responseInfo: ResponseInfo