+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/JapanSearchJavaExampleActivity.java
@@ -45,7 +45,7 @@ public class JapanSearchJavaExampleActivity extends AppCompatActivity {
         }
 
         @Override
-        public void onCategoryResult(@NonNull SearchSuggestion suggestion, @NonNull List<SearchResult> results, @NonNull ResponseInfo responseInfo) {
+        public void onResults(@NonNull SearchSuggestion suggestion, @NonNull List<SearchResult> results, @NonNull ResponseInfo responseInfo) {
             Log.i("SearchApiExample", "Category search results: " + results);
         }