+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/JapanSearchKotlinExampleActivity.kt
@@ -40,7 +40,7 @@ class JapanSearchKotlinExampleActivity : Activity() {
             Log.i("SearchApiExample", "Search result: $result")
         }
 
-        override fun onCategoryResult(
+        override fun onResults(
             suggestion: SearchSuggestion,
             results: List<SearchResult>,
             responseInfo: ResponseInfo